fix(chord_composer): more safely handle the placeholder ZWSP
sometimes Rime outputs an invalid character, invisible among text or being
visualized as diamond with question mark in macOS. I observed this bug often
when I was using combo pinyin but still not sure if this code has to do with
the issue. just mark the placeholder segment "phony" to prevent accidentally
commiting the ZWSP character.